Remove a warning.
[git-p4: depot-paths = "//depot/qt/main/": change = 131206]

+++ src/core/kernel/qcorevariant.cpp
@@ -1358,7 +1358,7 @@ static const char* const type_map[ntypes] =
 */
 const char *QCoreVariant::typeToName(Type typ)
 {
-    if (typ >= ntypes)
+    if ((int)typ >= ntypes)
         return 0;
     return type_map[typ];
 }
